+++ Source/Mockolate/Parameters/IParametersMatch.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace Mockolate.Parameters;
 
 /// <summary>
@@ -8,5 +10,5 @@ public interface IParametersMatch
 	/// <summary>
 	///     Checks if the <paramref name="values" /> match the expectations.
 	/// </summary>
-	bool Matches(object?[] values);
+	bool Matches(ReadOnlySpan<object?> values);
 }